Move Twitter stream logic to separate file
